[#39] :bug: fix: sns login component에서 로그인 처리하면 token 인식 못하는 문제 해결
